requirements.txt
duckdb-1.5.2
pandas-3.0.2